impl. pr suggestions.
