+++ packages/ui/src/utils/keyboard.ts
@@ -1,0 +1,2 @@
+// 'Esc' -> IE/Edge specific value
+export const escKeyFilterPredicate = (event: KeyboardEvent) => event.key === 'Escape' || event.key === 'Esc'
